+++ fire-group-project/src/App.js
@@ -1,9 +1,6 @@
 
 import './App.css';
-import fetchData from './services/constants';
 import {NavLink, Route, Routes} from 'react-router-dom'
-import { useEffect, useState } from 'react'
-import Input from './components/Input';
 import Home from './components/Home';
 import AboutUs from './components/AboutUs';
 import Zodiac from './components/Zodiac';
@@ -12,10 +9,6 @@ import Zodiac from './components/Zodiac';
 
 
 function App() {
-
-fetchData()
-
-
   return (
     <div className="App">
       <nav className="buttons">
@@ -25,7 +18,7 @@ fetchData()
           <NavLink to="/horoscope">Daily Horoscope</NavLink>        
           <NavLink to="/zodiac">Zodiac</NavLink>
           <NavLink to="/">Home</NavLink>
-          
+
       </nav>
