+++ fire-group-project/src/components/Button.js
@@ -1,12 +1,12 @@
-
+import TimeFrame from "./TimeFrame"
 
 function Button({onClick, onChange}){
 
     return(
         <div>
-        <input type="text" placeholder="Enter Horoscope" onChange={onChange}/>
-        <button type="submit" onClick={onClick} >Search</button>
+        
         </div>
+
     )
 
 }
