--- fire-group-project/src/components/Input.js
@@ -1,13 +1,0 @@
-
-
-function Input({}){
-
-    return(
-        <div>
-       
-        </div>
-    )
-
-}
-
-export default Input
