+++ fire-group-project/src/components/TimeFrame.js
@@ -1,0 +1,11 @@
+
+
+function TimeFrame(){
+    return(
+        <div>
+            
+        </div>
+    )
+}
+
+export default TimeFrame
