+++ fire-group-project/src/services/constants.js
@@ -2,26 +2,18 @@ import axios from 'axios'
 
 
 
-async function fetchData(input, day){
-    const DEFAULT_URL = `https://aztro.sameerkumar.website/?sign=aries&day=today`
-    const BASE_URL= `https://aztro.sameerkumar.website?sign=${input}`
+async function fetchData(sign, timeframe){
+    const DEFAULT_URL = `https://aztro.sameerkumar.website/?sign=${sign}&day=${timeframe}`
+    const BASE_URL= `https://aztro.sameerkumar.website?sign=`
     // &day=${day}
 
- try{
-        
-    if(!input){
-        let response  = await axios.post(`${DEFAULT_URL}`)
-        let result = response
-        console.log(result)
-        return result   
-    }
-    else{
-     let response  = await axios.post(`${BASE_URL}`) 
-       let result = response.data
-       return result   
-    }
+    // const signs = ['aries','taurus','gemini','cancer','leo','virgo','libra','scorpio','sagittarius','capricorn','aquarius','pisces']
+
+    try{
+        const response = await axios.post(DEFAULT_URL)
+        console.log(response);
     
-}
+    }
     catch (error) {
         console.log(error)
     }
